+++ backend/src/pytorch/InterpolateArchs/GMFSS/GMFSS.py
@@ -7,13 +7,8 @@ from .FeatureNet import FeatureNet
 from .gmflow.gmflow import GMFlow
 from .MetricNet import MetricNet
 from .FusionNet_u import GridNet
-from ....constants import HAS_SYSTEM_CUDA
+from ....utils.Util import CudaChecker
 from ..DetectInterpolateArch import ArchDetect
-
-if HAS_SYSTEM_CUDA:
-    from ..util.softsplat_cupy import softsplat
-else:
-    from ..util.softsplat_torch import softsplat
 
 
 class GMFSS:
@@ -44,6 +39,10 @@ class GMFSS:
         tmp = max(_pad, int(_pad / self.scale))
         self.pw = math.ceil(self.width / tmp) * tmp
         self.ph = math.ceil(self.height / tmp) * tmp
+        if CudaChecker.checkForCUDA():
+            from ..util.softsplat_cupy import softsplat
+        else:
+            from ..util.softsplat_torch import softsplat
         self.warp = softsplat
 
         combined_state_dict = torch.load(model_path, map_location="cpu")
@@ -61,7 +60,7 @@ class GMFSS:
         # model unspecific setup
 
         self.ifnet = IFNet(ensemble=ensemble).to(dtype=dtype, device=device)
-        self.flownet = GMFlow().to(dtype=torch.float, device=device)
+        self.flownet = GMFlow().to(dtype=dtype, device=device)
         self.metricnet = MetricNet().to(dtype=dtype, device=device)
         self.feat_ext = FeatureNet().to(dtype=dtype, device=device)
         self.fusionnet = GridNet().to(dtype=dtype, device=device)
@@ -114,9 +113,9 @@ class GMFSS:
             import gc
 
             gc.collect()
-            torch.cuda.empty_cache()
-            torch.cuda.reset_max_memory_allocated()
-            torch.cuda.reset_max_memory_cached()
+            #torch.cuda.empty_cache()
+            #torch.cuda.reset_max_memory_allocated()
+            #torch.cuda.reset_max_memory_cached()
             self.ifnet = trtHandler.load_engine("IFNet.engine")
             self.feat_ext = trtHandler.load_engine("Feat.engine")
             self.flownet = trtHandler.load_engine("Flownet.engine")
@@ -138,8 +137,8 @@ class GMFSS:
             imgf0 = img0
             imgf1 = img1
         if self.flow01 is None:
-            self.flow01 = self.flownet(imgf0.float(), imgf1.float()).to(dtype=self.dtype)
-            self.flow10 = self.flownet(imgf1.float(), imgf0.float()).to(dtype=self.dtype)
+            self.flow01 = self.flownet(imgf0, imgf1)
+            self.flow10 = self.flownet(imgf1, imgf0)
         if self.scale != 1.0:
             self.flow01 = (
                 F.interpolate(