+++ backend/src/pytorch/InterpolateIFRNET.py
@@ -44,6 +44,7 @@ class InterpolateIFRNetTorch(BaseInterpolate):
         self.ensemble = ensemble
         self.hdr_mode = hdr_mode # used in base interpolate class (ik inheritance is bad leave me alone)
         self.UHDMode = UHDMode
+        self.gpu_id = gpu_id
         self.CompareNet = None
         self.max_timestep = max_timestep
         if UHDMode:
@@ -82,8 +83,8 @@ class InterpolateIFRNetTorch(BaseInterpolate):
 
     @torch.inference_mode()
     def _load(self):
-        self.stream = self.torchUtils.init_stream()
-        self.prepareStream = self.torchUtils.init_stream()
+        self.stream = self.torchUtils.init_stream(self.gpu_id)
+        self.prepareStream = self.torchUtils.init_stream(self.gpu_id)
         with self.torchUtils.run_stream(self.prepareStream):  # type: ignore
             
             from .InterpolateArchs.IFRNET.IFRNet import IFRNet