+++ backend/src/pytorch/UpscaleTorch.py
@@ -3,6 +3,7 @@ import math
 
 import gc
 from .TorchUtils import TorchUtils
+from .spandrel.UpscaleModelWrapper import UpscaleModelWrapper
 import torch as torch
 import torch.nn.functional as F
 import sys
@@ -89,9 +90,9 @@ class UpscalePytorch:
 
     ):
         self.torchUtils = TorchUtils(width=width, height=height,hdr_mode=hdr_mode,device_type=device)  
-        device = self.torchUtils.handle_device(device, gpu_id)
-        self.tile_pad = tile_pad
+        device = self.torchUtils.handle_device(device, gpu_id=gpu_id)
         self.dtype = self.torchUtils.handle_precision(precision)
+        self.tile_pad = tile_pad
         self.device = device
         self.videoWidth = width
         self.videoHeight = height
@@ -99,7 +100,6 @@ class UpscalePytorch:
         self.tile = [self.tilesize, self.tilesize]
         self.modelPath = modelPath
         self.backend = backend
-        
       
         self.trt_workspace_size = trt_workspace_size
         self.trt_optimization_level = trt_optimization_level
@@ -111,16 +111,11 @@ class UpscalePytorch:
         self.trt_static_shape = trt_static_shape
 
         # streams
-        self.stream = self.torchUtils.init_stream()
-        self.f2tstream = self.torchUtils.init_stream()  
-        self.prepareStream = self.torchUtils.init_stream()
-        self.convertStream = self.torchUtils.init_stream()
+        self.stream = self.torchUtils.init_stream(gpu_id=gpu_id)
+        self.f2tstream = self.torchUtils.init_stream(gpu_id=gpu_id)
+        self.prepareStream = self.torchUtils.init_stream(gpu_id=gpu_id)
+        self.convertStream = self.torchUtils.init_stream(gpu_id=gpu_id)
         self._load()
-
-
-    # add prores
-    # add janai v2
-    # add bhi light model
 
     @torch.inference_mode()
     def _load(self):
@@ -140,9 +135,13 @@ class UpscalePytorch:
 
         
         with self.torchUtils.run_stream(self.prepareStream):
-            self.set_self_model(backend="pytorch")
+            self.upscale_model_wrapper = UpscaleModelWrapper(
+                model_path=self.modelPath,
+                device=self.device,
+                precision=self.dtype,
+            )
 
-            match self.scale:
+            match self.upscale_model_wrapper.get_scale():
                 case 1:
                     modulo = 4
                 case 2:
@@ -223,24 +222,9 @@ class UpscalePytorch:
                         dim_width = _width * modulo
                         dynamic_shapes = {"x": {2: dim_height, 3: dim_width}}
 
-                        
-
-                    # inference and get re-load state dict due to issue with span.
-                    try:
-                        model = self.model
-                        model(inputs[0])
-                        self.model.load_state_dict(model.state_dict())
-                        output = model(inputs[0])
-                        del model
-                        torch.cuda.empty_cache()
-                    except Exception as e:
-                       print("Test inf failed")
-
-                    
-
                     try:
                         trt_engine = trtHandler.build_engine(
-                            self.model,
+                            self.upscale_model_wrapper.get_model(),
                             self.dtype,
                             self.device,
                             example_inputs=inputs,
@@ -266,7 +250,7 @@ class UpscalePytorch:
                             else:
 
                                 trt_engine = trtHandler.build_engine(
-                                    self.model,
+                                    self.upscale_model_wrapper.get_model(),
                                     self.dtype,
                                     self.device,
                                     example_inputs=inputs,
@@ -283,16 +267,15 @@ class UpscalePytorch:
                             raise RuntimeError(
                                 f"Failed to build TensorRT engine: {e}\n"
                             )
-                self.set_self_model(backend="tensorrt", trt_engine_name=self.trt_engine_name)
-
-                
+                model = trtHandler.load_engine(trt_engine_name=self.trt_engine_name)
+                self.upscale_model_wrapper.load_model(model)
 
         self.torchUtils.clear_cache()
         self.torchUtils.sync_all_streams()
 
     @torch.inference_mode()
     def hotUnload(self):
-        self.model = None
+        self.upscale_model_wrapper = None
         gc.collect()
         self.torchUtils.clear_cache()
         if HAS_PYTORCH_CUDA:
@@ -302,67 +285,18 @@ class UpscalePytorch:
     @torch.inference_mode()
     def hotReload(self):
         self._load()
-
-    @torch.inference_mode()
-    def set_self_model(self, backend="pytorch", trt_engine_name=None):
-        torch.cuda.empty_cache()
-        if backend == "tensorrt":
-            from .TensorRTHandler import TorchTensorRTHandler
-            trtHandler = TorchTensorRTHandler(model_parent_path=os.path.dirname(self.modelPath),)
-            self.model = trtHandler.load_engine(trt_engine_name=trt_engine_name)
-        else:
-            self.model = self.loadModel(
-                modelPath=self.modelPath, device=self.device, dtype=self.dtype
-            )
-
-    @torch.inference_mode()
-    def loadModel(
-        self, modelPath: str, dtype: torch.dtype = torch.float32, device: str = "cuda"
-    ) -> torch.nn.Module:
-        try:
-            from .spandrel import ModelLoader, ImageModelDescriptor, UnsupportedModelError
-        except ImportError:
-            # spandrel will import like this if its a submodule
-            from .spandrel.libs.spandrel.spandrel import ModelLoader, ImageModelDescriptor, UnsupportedModelError
-        try:
-            model = ModelLoader().load_from_file(modelPath)
-            assert isinstance(model, ImageModelDescriptor)
-            self.model = model
-            # get model attributes
-            
-        except (UnsupportedModelError) as e:
-            from .VSRArchs.AnimeSR import AnimeSRArch
-            from .VSRArchs.vsr_inference_helper import VSRInferenceHelper
-            model = AnimeSRArch()
-            self.inference = VSRInferenceHelper(model)
-
-        self.scale = model.scale
-        model = model.model
-        
-        model.load_state_dict(model.state_dict(), assign=True)
-        model.eval().to(self.device, dtype=self.dtype)
-        try:
-            example_input = torch.zeros((1, 3, 64, 64), device=self.device, dtype=self.dtype)
-            model(example_input)
-        except Exception as e:
-            print("Error occured during model validation, falling back to float32 dtype.\n")
-            log(str(e))
-            model = model.to(self.device, dtype=torch.float32)
-            
-        return model
-
     
     @torch.inference_mode()
     def __call__(self, image: bytes) -> torch.Tensor:
         image = self.torchUtils.frame_to_tensor(image, self.f2tstream, self.device, self.dtype)
         with self.torchUtils.run_stream(self.stream), torch.amp.autocast(
             enabled=self.dtype == torch.float16,device_type="cuda"):
-            while self.model is None:
+            while self.upscale_model_wrapper is None:
                 sleep(1)
             if self.tilesize == 0:
-                
-                output = self.model(image)
-                
+
+                output = self.upscale_model_wrapper(image)
+
             else:
                 output = self.renderTiledImage(image)
             
@@ -375,7 +309,7 @@ class UpscalePytorch:
         return output
  
     def getScale(self):
-        return self.scale
+        return self.upscale_model_wrapper.get_scale()
 
     @torch.inference_mode()
     def renderTiledImage(
@@ -431,7 +365,7 @@ class UpscalePytorch:
                 )
 
                 # process tile
-                output_tile = self.model(
+                output_tile = self.upscale_model_wrapper(
                     input_tile
                 )