+++ backend/src/utils/BackendDetect.py
@@ -42,7 +42,7 @@ class BackendDetect:
 
     def __get_pytorch_device(self):
         if "cu" in self.__torch.__version__: return "cuda" 
-        if "rocm" in self.__torch.__version__: return "cuda"
+        if "rocm" in self.__torch.__version__: return "rocm"
         if self.__torch.xpu.is_available(): return "xpu"
         if self.__torch.backends.mps.is_available(): return "mps"
         return "CPU"
@@ -59,7 +59,7 @@ class BackendDetect:
         """
 
         try:
-            x = self.__torch.tensor([1.0], dtype=self.__torch.float16).to(device=self.pytorch_device)
+            x = self.__torch.tensor([1.0], dtype=self.__torch.float16).to(device="cuda" if self.pytorch_device == "rocm" else self.pytorch_device)
             return True
         except Exception as e:
             log(str(e))
@@ -78,6 +78,7 @@ class BackendDetect:
             torch_cmd_dict = {
             "cuda": self.__torch.cuda,
             "xpu": self.__torch.xpu,
+            "rocm": self.__torch.cuda,  
             }
 
             torch_cmd = torch_cmd_dict[self.pytorch_device]